refactor(electron): extract codebase-indexing queries out of db/queries.ts
Continue the db/queries.ts split: move the codebase semantic-indexing queries
(file/symbol/relation CRUD, codebaseScope, search/definition/relations,
getCodebaseOverview/Graph/ModuleGraph) into electron/db/codebase-queries.ts
(~410 lines), depending only on the Database handle, ts() and path. queries.ts
re-exports it (export * from "./codebase-queries") so every ../db/queries /
./queries import (lib/codebase-index.ts, mcp/tools/codebase.ts, ipc/agent.ts,
components/agent/ModuleMap.tsx) is unchanged. Dropped the now-unused path import
from queries.ts. queries.ts 1671 -> 1267 (1954 -> 1267 across both slices).

No Database constructed in the new file (caller's handle), so it stays safe for
both the Electron main process and the bundled MCP server.

Verified: type-check:all + eslint clean; queries/codebase-index/grep-vs-semantic
tests (69) + full suite (1128) green; all four esbuild bundles compile.
