+++ changelogs/v2.4.14.md
@@ -8,3 +8,4 @@
 ### Under the hood
 
 - Split the note + task embedding queries out of the 1,950-line database-query module into their own file. Purely internal reorganisation — every existing import keeps working via a re-export; no behaviour change.
+- Also split the codebase semantic-indexing queries (files, symbols, relations, and the Architecture/Module-Map rollups) into their own file, the same way. Purely internal; no behaviour change.
